Revert "Use AlertDialog for MediaRouteControllerDialog"
This reverts commit 355e861498fd44c5dcfb2f0c0538dbbeaa6ed4ca [formerly 001d51496d062789355a91ce9365cd0cfeac6925].

Background: To just inherit AlertDialog, the content view should include
a title as we do in support library (AlertDialog uses NO_TITLE feature),
but up-streaming support library implementation to the framework at this
point might cause more issues. Verified that the narrow dialog issue
(b/22044600) does not happen in the framework implementation regardless of
whether it uses AlertDialog or not.

Bug: 22286869
Change-Id: Ic2554cc9e683beff29d1deee91945c1dace83ab1

Former-commit-id: 7980d10d7c1ac0b5301eb8efaa18657201074910