Docs: Revisions to top DAC landing page and several Preview-page revisions.
Top landing page now has a link direct to Preview 2.
The "Program Overview" now explicitly mentions that Preview 2 is available,
and is updated so that it no longer refers to Previews 1 and 2 in the future
tense. "Support" page is renamed to "Support and Release Notes."
TOC revised to reflect new page name. Direct links to Preview 2 are added
to the respective languages' Preview landing pages.

Change-Id: Ic463b123b25224b7064b4771faa79bdfe63e9ff5

Former-commit-id: 6fb22b4e42c18c91238cc687f84e4fe5f2950c52