Smoothly animate notification pane if opened via keyboard.
Animation would be slow and then 'jump' to the finished state if opened
with Meta-n. This was caused by the panel view calculating the height of
the notification shade based on a pre-expansion value.

The fix uses instantExpand instead of expand (which we no longer need);
the former sets up proper state such that the height of the window is
properly determined prior to the fling height being determined. We just
add an animate option to instantExpand so we can reliably animate open
from a closed starting state.

Bug: 22359801

Change-Id: I8a3a96041a65299ca985010d43bd4aa12827c8fa

Former-commit-id: 8d3fea15b80ac4fab7abe7a7a07c6368166de5e1