Fix input pause without resume.
Bug 22455206

Previously, when an exit activity transition was created, the input
would be paused. This worked fine as long as the transition was
run. However, sometimes that transition wasn't run and this would
cause the input to fail to be started again.

This fix moves the input pause to when the transition is started.

Change-Id: I738d5471f7932f00b50897f87a8f8a71ecbc57e2

Former-commit-id: 80141d1c8b799cf7669a29b8837037cf3934ddb9