Make additional shared elements a version-checked feature.
Bug 22505481

ag/684544 added a feature to allow shared elements that weren't
shared into an Activity to be shared back. However, if you are
targeting an older version of the SDK, you may get an unexpected
shared element back. This change in behavior has been locked
behind a target version check.

Change-Id: I7162e24f3b14fedd6b308e89e9d04ac67660f7d6

Former-commit-id: 2c256a04ae551780c733eea17995b4b2ce69eecd