Always create nodes for views backed by an adapter.
Some views are backed by adapter and if the adapter changed
but views are not updated we were not reporting the views.
This is not correct as the accessibility layer should always
access whatever is on the screen regardless if we know it
will change in the next layout pass.

bug:20920903

Change-Id: I5851c886848e7b8e59b76419c22124790d7e6f05

Former-commit-id: 457ec9e66810ac5351399ca12897ba0676e607d4