docs: Updated the list of auto partners, added links
Updated the list of auto partners to the full, current list of 38.
Added logos for partners that were missing; made all logos into
hyperlinks to companies' world-wide sites.

NB: I did my best to find the world-wide sites for the automakers,
but I'm not 100% sure about all of them. In many cases the www.*.com
address redirected to a US-specific site, but I'm assuming that's
because my browser is in the US--I assume that viewers in other
countries will see their country's site or an international version.
(In a few cases I was able to find a specific URL for the worldwide
site.)

See first comment for doc stage location.

bug: 22619956
Change-Id: Ib55222deefadba59c0ba8ca56a56cfba31372545

Former-commit-id: dc5bb53d78e7010ff0bd92ae933cf390aa68ecd7