Fixed issue with not finding existing activity for background user
We currently go through the list of existing activities to find
a match for the launching intent so we can re-use the activity
record if one already exist. However, we exit the search early
once we run across an activity record that doesn't belong to the
current foreground user. This will cause us to create duplicate
activity records if the launching intent is for a backround
user and an activity record already exist.

Based on https://android-review.googlesource.com/#/c/159131

Bug: 22564256
Change-Id: I4b6d94059c11fd2e621e65c8ec2c99427c15b246

Former-commit-id: 25073dd4a8e8efcfe4b0b728ebb98fb7326434c0