Work on issue #22303510: Additional permissions aren't properly...
...disabled after toggling them off

Keep track of whether a permission that has been declared by an app
was able to actually be installed in the system, along with an API
to find this information so that system UI can tell whether that
permission is of interest.

Also clean up some of the permission debug output.

Change-Id: If4541bedb857789b255bb18f03cad155dcda0b95

Former-commit-id: cfbfafe1b9ca2fd135a4fb6b528b3829830803bf