Partially fix that beginning of animation is skipped
When waiting for all the windows that belong to an activity, we
skipped the main window, in case it didn't had a surface yet. This
was a problem because with SurfaceViews: They set it's visibility
extremely early in the app visibility change cycle. Then, they don't
block with the callback to SurfaceHolder.Callback2.surfaceRedrawNeeded,
so window manager thinks it has drawn the first frame already. We don't
even have a surface for the main window yet, but we start the app
transition already because we think the only interesting window for this
app token is the SurfaceView, which has supposedly already drawn.

Bug: 22207948
Change-Id: I2ddb274000fa08866e193f4fd8db3788fc55e0f1

Former-commit-id: ac73f1efea0cb23881571e18e87ffc7f388af7ad