UsbDeviceManager: fix b/21429947 regression
As discussed in b/21429947 (commit
5a5c8aa123ddbee7d4e16a2abcb65414be247652 [formerly 674019065bceb4150190bfb1aa63cda9de0a8560]), MTP must always be
enabled, even if access to the underlying MTP data is disabled.
Otherwise, Android will not enumerate on the USB bus, and won't
receive notifications from the kernel about USB state changes. This
effectively prevents using MTP functionality on user builds, or
on userdebug/eng builds with adb turned off.

Always ensure that MTP is the default driver mode.

Get rid of one use of the persistent property. The persistent property
was already pulled from a number of devices, and as explained in
commit 19ea18f324948fb1c9d0ac1cbdda0d4a3ffb39f0 [formerly fcf10f7c12cb3107bdfedce6f76a8c866d154f3c], the intent was that
the persistent property would only hold the persistent adb state.

Bug: 22447614
Bug: 21429947
Change-Id: I8b3690a1bafb7cea0d5a69d73c1065c7fc64c653

Former-commit-id: b210026e3d5c955628ca8b8b9191ade08891e9ef