Move updateMatrix() call to animateCommon
Bug: 22667315

HWUI internally always assumes that RenderNode->properties()
has a fully-baked matrix. However, RenderThread animations violate
this and will perform an update of that matrix after they run. This
was missing in the fallback scenario where an animation was
running on a View that is detached from the tree.

If a RenderThread animation is started on an attached view, it becomes
detached, and is then re-attached it will crash in the staging
properties push. The fix for this is to move the call to updateMatrix
from animate() to animateCommon() so that the animateNoDamage()
path will similarly perform the necessary matrix update.

Change-Id: I4c899a6f68659342a6515f1d7f20d3c4af65caf0

Former-commit-id: 49dec430e8a38943c5e934c1e31b724bf53c47c4