LayoutLib: Fix Status/Navigation Bar color.
If the color value for status bar or navigation bar is declared directly
in the theme (i.e. doesn't reference a color resource via @color/foo),
the ResourceType for the attribute is not assigned by the IDE. LayoutLib
used to ignore resources that were not of type color. This change fixes
the issue by also checking the resources without a type.

Change-Id: I94735ec225415282db06ab9db5c3233ad89c052f

Former-commit-id: e3f9834ebd8a6e4ee155c8e4c9456dfc98e8b832