Don't animate burn-in return when it wasn't applied.
The burn-in protection return animation is performed based on the last
offset that was used, but it doesn't mean that such offset is currently
applied. Right after entering ambient mode no offset is applied, so if
we start a return animation using the last applied offset, it will cause
a shake. The animation will first shift from 0, 0 offset to the last
offset and then return again to 0, 0. This CL makes it not apply the
animation in that case.

Bug: 22519478
Change-Id: I5cb750b56e9715c6d9389136071a5e0fbafb047b

Former-commit-id: 0b388d4a368e8503411450504c6e3d39952bce16