Workaround View.post issues to fix runtime crash
Using View.post was really dangerous because when the view wasn't
attached, it got posted on the run queue of the *calling* thread.
However, that run queue was never executed until power down, and
then it was executed from the PowerManagerService thread, because
that was the calling thread when we posted it. Work around this by
using a solid Handler.

Bug: 22820787
Change-Id: Id60e49e859558993256fae0403236f2e4b6f1075

Former-commit-id: 4a0e791cc0a71df2b0c6bae6e92a5f2100fb5841