Make getInt in BridgeTypedArray accept empty String.
The XML editor already warns about the empty strings in places where an
int value is required. There's no need to show another warning for it in
the rendering panel. Also, the rendering might have also failed when the
empty string is encountered. Try an continue the rendering, because no
rendering state is really bad.

Change-Id: I85106f37e4462f237e85f0b065b4ce8a4bfabf4d

Former-commit-id: 0437cac76c24dc6cac7b530e6d80f789565550ec