Fix issue #22843150: fragment transaction crash from onRequestPermisssionResult
Add new Activity callback to tell it when its saved state has
been invalidated.

The problem is that delivering the permission result does not go
through a path where the compatibility code can see it first to
mark its fragment manager as no longer having saved state.  So this
new callback gives it a place to do that.

Change-Id: I5a4a185d9c746bae1afb5c588aba82c8daccf079

Former-commit-id: 6bdd3a155fe380735d4108ddcc50b156c62406d0