Using workspace to dertermine decor to use
This patch uses the assigned stack to determine if
a non client decor needs to be added or not.

Since the visibility of the non client decor is
dependent on the used workspace, and an application
can transfer to a different workspace at any time,
the visibility might also change at any time.
As such this patch will also create the nc decor
for every window which might need it later in time.

BUG: 22984908
Change-Id: Ic7c0b3fd93d021bce882a5345738ceb3f608a3b9

Former-commit-id: f4dd71a2657d23d12d5fe6e5e1273aab65d80ca4