Fixed focus adjustment issue with freeform stack.
The activity focus was swifting from the freeform stack to
the home stack even though there were still visible tasks
in the stack. This was because the rules of handling focus
adjustment for the fullscreen stack was been applied to
freeform stack. We now keep the focus on the freeform
stack as long as there is a running activity in the stack
that we can adjust focus to.

Bug: 23035924
Change-Id: I8f81299c74b54e751400cfa265243018e69695d2

Former-commit-id: 61803ee5137d4235f17de9e78d936af4a44c2bff