Don't call authenticate immediately after getting valid fingerprint
After gettin a callback onAuthenticationSucceded we set the
fingerprint listening state to false. However, when waking up, we
immediately started listening again because the state was false.
Protect against that by only calling authenticate only once, except
when the unlock doesn't go through because unlocking with fingerprint
is not allowed.

Also fixes some animation "jank" because of the state messup.

Bug: 23304421
Change-Id: Ic83ac0f1590dd4f8017bb55dca9e19a60cfdf99f

Former-commit-id: 71448a70036e8f95a4f7f3b0f6252bd3344976bb