Make wallpaper visible in freeform mode.
The system requires at less one window to be
a wallpaper target in order to make the wallpaper visible.
For example the home and recents activities are both
wallpaper targets. For the case of the freeform workspace
there is no activity/window that is a wallpaper target.
We now make the wallpaper window target itself when in freeform
mode so we don't need an additional window to make it visible.

Change-Id: I6553c0fcec770e931c08bed8ef5c533b47afe1b8

Former-commit-id: 21fdd916a77e0ec64ca1c9b7fcb454283d6f58c3