Using initial activity layout to position the launching activity.
The activity can be positioned in the center or in one of the corners.
From there it shifts its position and tries to find a spot where it
doesn't collide with other activities.

The CL also includes a few fixes necessary to pipe the information about
the initial layout through the system.

Change-Id: I2aaf5b6d20044aafec713b7bd4193b05cfbd16f3

Former-commit-id: 9b1ce52f254b4d9c17ebf437f19f45603d3ad5b2