Remove window from mWaitingForDrawn when policy hidden
When doing a wake-and-unlock sequence and the app underneath hides
the status bar, we are dismissing Keyguard, the app gets focus, the
app hides the status bar BUT we haven't finished drawn yet, so we
are waiting until the timeout because w.hasDrawnLw() always returns
false. Remove window from pending list if it gets policy hidden.

Bug: 23431343
Change-Id: I27e8c6d4efb435a4d3b632a4a37c0de5b021c2e7

Former-commit-id: 414985c422664c7f8d5dec5562b502907b965876