Camera2: Remove filtering for non-BURST_CAPTURE-capable devices
Devices that do not support the BURST_CAPTURE capability do not
get the split resolution lists. However, one spot in the stream
configuration map enumeration did not guard against this properly.

As a result, the category of devices that are not-BURST-CAPABLE and
do list resolutions that operate at a 'slow' (<= 20fps) rates cause
an assert to fire from an apparently-inconsistent size list.

Guard the filtering correctly, to correctly support such
devices.

Bug: 23489536
Change-Id: Ib026f6555c69b7fac0cd1ea5027d598644cd8701

Former-commit-id: e3f54839770066bbdd2886c962f5c72bd40f7d20