Test ValueAnimator with custom timing pulse provider
This CL depends on ag/749611

With the custom timing pulse provider, we are free from the chaos that could
happen on UI thread and any delay that comes as a result. Therefore we can
count on more accurate timing for each frame, which enables us to test things
such as AnimatorUpdateListener, among other things.

Note that there will still be ~2ms delays in addition to the frame interval
in between frames, due to the MessageQueue processing time.

Change-Id: I6de0aa06125d1ecfae788f5dd22536a5fa52e651

Former-commit-id: 2822a426ab48f073088e185a94c9d8ee41d7b383