Reformat only KR numbers with leading "+82".
Looks like the phonenumber util can figure out the country code of a
phone number to format from the localization data even if it is not
explicitly present in the phone number string. So, adding an explicit
check to make sure that the country code(+82) was fetched from the
original number string which was the intention of the fix for
b/22862845.

BUG: 23661668
Change-Id: I26ba03d759e0ec35a195ba51bf60d09eff8de090

Former-commit-id: a9acb4c2b55fe1bf4f5eeb96971e46697eaeffe9