Keep local run queue in View to avoid posting to wrong thread
Previously any call to post() while a view was detached would stash the
runnable on a thread-local RunQueue. If the post() call happened off the
UI thread to which the view would later be attached, this would result
in the runnable either failing to run or running on some other UI thread.

Bug: 22828132
Change-Id: I5289ee91ea4b56c45f003a387a9f7f0e6dc050f9

Former-commit-id: bea0c7daa6611d8b96e1271f8854f500a87342fc