Make the vertical navbar 48dp like the horizontal one.
When the navigation bar was originally introduced to phones
(in change Ic613a335) we wanted it to stick to the same spot
on the display so that it felt as much as possible like a
physical button array; pressing the same fingertip-sized
spot on the glass should always invoke BACK, etc. This meant
flipping the nav bar to a vertical orientation when the
phone was in landscape, and then juggling around the window
insets and other system windows to make room for it.

For reasons that are now lost to time, in that original
implementation we made the vertical navigation bar narrower:
42dp (versus 48dp for the horizontal navigation bar, which
incidentally is always horizontal on tablet-type devices).
Nobody really noticed (except app developers looking to
hardcode this value instead of just using fitSystemWindows
or the new WindowInsets).

Here we finally make the navigation bars match perfectly in
portrait and landscape.

Bug: 23724209
Change-Id: I861be84b41c6a227d269469686c8c66a32029f1d

Former-commit-id: c263a086637f0a140302bb5e686bbce4e0a6fb18