Ensure entering and exiting views don't match in Fragment Transitions.
Bug 23693301

An enter or exit transition targeting an ID, name, or similar
would match by the given criteria. This disreguarded that the
exiting view should not match against the entering views, even
if they have the matching criteria.

This excludes the exiting and shared elements from participating in
the enter transition and the enter and shared elements from
participating in the exit transition.

Change-Id: I4007699d2546416008f7a9fcfdb87ff504316577

Former-commit-id: 61f83319292699658bcf4e0e0a4244832e299c84