Fix scrolling of QS panel when no notification is present.
After boot when status bar donot have any notification, the
overscroll is rubberbanded , changes have been done to fix
this and make the behaviour consistent with other scenarios.

In NotificationStackScrollLayout when there is no childview
present at the touch coordinate, initialization of some
variable was not done which was resulting into overscroll
being rubberbanded.

Usecase of the issue:
1. Reboot the phone.
2. Make sure that no notification is present
3. Pull the status bar.
4. Now pull the "No notifactions" view ,the overscroll
   is rubberbanded
5. Get some notifications, pull the status bar and then
   the notifications, the overscroll is not rubberbanded
6. Remove all the notifications, now pull the statusbar
   and then pull "No notifications" view the overscroll is not
   rubberbanded.

Change-Id: I06d4f0e3d6ed980744be21b9fd7ed307dca69ca3
Signed-off-by: Jayasri bhattacharyya <jayabhatt@codeaurora.org>

Former-commit-id: 5e55c8974c4443dd08d8aa1d6082dca6d940393f