Properly resize freeform stack when docked stack is resized.
We were previously setting the task bounds for freefrom tasks to the
stack bounds when the stack is resized. So, all the tasks in the stack
are sized to the stack bounds vs. trying to fit in the stack bounds
while maintaining their size.
Also, we only want to apply fit to bounds on a task bound when we are
resizing the stack. So, moved method from TaskRecord to
ActivityStackSupervisor.

Change-Id: Ia6e046d6f2c3b0d80b06530a3e07d6ed1bb9b79f

Former-commit-id: 3e3e66f418dfb8ddd30e561e9bbc73ccf9ba0ca2