Update SMP documentation.
The old version was seriously obsolete, primarily in that it ignored
C11 and C++11 atomics and the programming model underlying them.  As
a result it paid way too much attention to hardware characteristics,
which 0.001% of application programmers should really be aware of.
And some of those hardware descriptions were also obsolete.

This is a fairly complete rewrite.

Bug: 18523857

Change-Id: Icc14a390f74486193486c2ba07a86b05611e7d3c

Former-commit-id: 9e629adc4ae39b8ced5a0ff74d5a5049a44a7286