Making sure that after a rotation Shelf shows latest set of pinned apps.
2 copies of NavigationBarApps (for vertical and for horizontal orientation) share a single
static copy of NavigationBarsAppsModel.
If the active orientation's view made changes, they are not conveyed
to the inactive one, and if we rotate to show the other view, it will show
stale data.

The fix introduces an event generated by the model, which allows the invisible
shelf to get updated when the visible one makes changes.

Bug: 20024603
Change-Id: I0d749dbac700857b081bce8a5a232f24ba271b25

Former-commit-id: 4dc0c15a6ac0564b2e7d9d8b957795a373d05907