Unbreak build, bring ConnectivityService in sync with mnc-dr-dev
The version of ConnectivityService.java in mnc-vt-dev doesn't
compile due to the presence of two similar-but-not-identical code
blocks in handleRegisterNetworkAgent. The history here is a
a little convoluted - this code was originally merged into
mnc-vt-dev, then cherry-picked into dr-dev, and from there
automerged into mnc-vt-dev, causing conflicts. This latest
breakage is likely due to the automerged not detecting a conflict
because the code block was subtly different.

Attempt to fix this once and for all by making the mnc-vt-dev
version of the file identical to the mnc-dr-dev version.

Change-Id: I270739b0be6f6358045700494a1b0f25f0b365a3

Former-commit-id: ba72ac498b773d2dc082a749eacd87b41937088a