Fix Notification settings upon installation after Restore.
When new package is installed, code incorrectly checked if any actions should
be taken in onPackagesChanged(), there was no check if this new package is
listed in :restored settings.

Rest of CL is related to building list of packages mentioned in :restored setting,
and to adding unified mechanism for reading/storing list of component
names to/from setting.

Bug 20764607

Change-Id: Id9b1843fe8d39f40751ff2eccbc8dd2a488ef5ac

Former-commit-id: 76c02686cd02efc3b21874919c54d5bbeb0f069f