Fixed issue with relaying-out sysui when resizing a freeform task.
WindowState.getTask() was returning the focusedApp task for windows
without an AppToken (e.g. status bar, nava bar, wallpaper). We then
check to see if the config associated with the task has changed and
tell the client window to resize if it as. The config would have
changed in this case since we are using the focus app task
(task we are currently resizing) to determine config changes for all
windows without appTokens. Changed the logic so this is no longer the
case.

Also, cleaned-up the use of getTask().

Bug: 23939846
Change-Id: I473671021282695d282a8a2b82d679722273ca09

Former-commit-id: 5a2f2cb3de21beaa0b760bb0be1928909012f81e