Adds an option to override RemoteViews transition.
This change adds a configuration option,
"config_overrideRemoteViewsActivityTransition", which is false by
default and true on watches.  If it's true, when a RemoteViews
is clicked, it will use a new attribute on the active window animation
style, "activityOpenRemoteViewsEnterAnimation", to determine
the entry transition to use, instead of the current hard-coded
default.

Bug: 23534750
Change-Id: I91e92804a663ab71a3e46c3430fa579b87cc7931

Former-commit-id: 1ed322b123b5aa8cd02c036e5e3393b6a79f559f