Don't resize task in current stack when moving it to a docked stack
The creation of a docked stack resizies all other stacks and thier
resizeable tasks. This will cause a task we are moving to a docked
stack to be initially resized in the current stack before it is
moved to the docked stack. We now disable resizeablility on the
task we are moving during the creation of a stack so we don't
resize it in the current stack.

Bug: 23978723
Change-Id: Ibb1c6d2a8493971a6cbccd782a378459fedcb162

Former-commit-id: eb8e56c8bbf9d14fd82e0a4e4972e7f9ad3479ef