Only place new task in focus stack if it is full screen stack.
The fullscreen stack is the only stack that can contain any task
regardless of if the task is resizeable or not. So, we don't want
to place any random new task in a non-fullscreen stack that has
focus as you can end up with issue like a non-resizeable task
ending up in the freeform stack and gets resized and the window
decor.

Bug: 23626353
Change-Id: Id2f151449f612b4f81ff85b42c375254871f6ec2

Former-commit-id: 13a9b16dc7a331687084dd3722b3134f0d928cb1