Fix black Keyguard
When a window has both the flag fullscreen and the dismiss Keyguard
flag, we end up in a state where we hide the status bar window but
all other windows, because mShowingLockscreen nevers gets set
correctly. Move it up so we always set it no matter whether the
status bar window was visible.

Bug: 22875357
Change-Id: I7953fe7100cc99fe8fb7424a9b311b4630426657

Former-commit-id: e1de9f674575df3b17e475c4f32dd93d33e596e3