Reduce the amount of allocations on Paint_Delegate
When the text properties of the Paint object are updated, or when .set()
is called, updateFontObject will be called. This method is currently at the
top of the methods for number of allocations.
This CL reduces the number of calls to updateFontObject by a 90%,
reducing the number of Font allocations.

Change-Id: I3e232f9e6e230da8ecdd19921faadf2d3f4ff763

Former-commit-id: 6dac0d4b7c55da32cb5deb9b589b1acb43cb536d