Call into PackageManager as mContext, not myUserId
ApplicationPackageManager instances can (and sometimes are) created as another
user from the one they run in. Example use: fetching the application icon for
an app installed as another user.

The occurrences that weren't doing this aren't ever called on a cross-user
instance so it should be safe enough to fix all of them.

Change-Id: Ia63f22942f35450e36a7b8da2bbf8eda360ef21a

Former-commit-id: 0e27c87af691bc53d320496bbffa58752608608d