Initialize DisplayContent base info when object is constructed
It is possible for an external process (e.g. sys-ui) to call into
WM to get the bounds of a stack in between the construction of the
DisplayContent object and SystemServer telling WM that the display
is ready. Incorrect bounds would be returned to the caller in this
case since the baseDisplayWidth/Height is only set when the displayReady
call is made.
We now initialize the base display information when the DisplayContent
object is created so any call to get the logical display rect of the
display will always return the right bounds. This should be okay as the
display manager is initialized way before window manager.

Bug: 24505761
Change-Id: Iebd688ad8f4c08556846fcebb543e58bd2c01414

Former-commit-id: fd04d8c8840c2a762e4f19c31310561263898621