Use visible window bounds to determine if a point is within a task
We need to use the visible frame on the window for any touch-related
tests. Can't use the task's bounds because the original task bounds
might be adjusted to fit the content frame. For example, the presence
of the IME adjusting the windows frames when the app window is the
IME target.

Bug: 24680002
Change-Id: Id73453788d56de879410027531ddefa46d929c67

Former-commit-id: 12cbd92c1dec4b93d2d292c3082f52d25cd6703b