Ensure SurfaceView requests WM relayout when size changes.
In the case of scaled SurfaceViews with a fixed size.
It is possible for the surface layout size to change
without a change to the surface size or position. In this
case we need to inform the window manager so it can update
frames and scaling. This has presented in Camera and required
workarounds such as changing the fixed surface size to trigger
an update (leading to visual artifacts).

Bug:23974105
Change-Id: I195aa0832907c32874b294101b56e198e853dce5
(cherry picked from commit f43ae5955f88c65ea5677c1b49b8eedbdb0049fa [formerly a8fe43f09c0ed2119e55139a482f7f12fbe7d725])

Former-commit-id: 57caeb5703de0c86be04f597d14e954987cd7955