Shared user retains permissions no longer used by any app on an app update
If an app in a shared user uses permission A and B and these are granted
to the shared user and now an app update is installed that only uses A,
the shared user still ratains the B grant. A shared user should have only
permissions declared as used by its currenlty installed apps.

bug:24736912

Change-Id: Idea6c06bdc236fd481a860cddb379e6ce660ee87

Former-commit-id: ff7b92fc02e3bf3a799716b396860d59e81886cb