Disassociate system windows from apps
With the original logic, if an app creates a system window, when the
user goes to home screen, the system window will be still there and
become unable to receive input events, because the system window will
be also changed to the stopped state with the app window, and the
current logic of ViewRootImpl forbid a stopped window receiving input
events.

This change prevents assigning the token of the app window to system
windows created by the app, so that when the app goes to the stopped
state, its system windows won't be affected (can still receive input
events).

This change is related to the following changes:
https://android.googlesource.com/platform/frameworks/base/+/ffc1cd6292e310477b5c6ed3aa47c9367aa65bfa [formerly a5d29971f8f4d463348e9de0f7af7e2e92a377a0]
https://android.googlesource.com/platform/frameworks/base/+/f9542daea17853625b0a7abda433553b66983dad [formerly 15ed2754a308546a54ab1061b4e365d6fe9b898d]
https://android.googlesource.com/platform/frameworks/base/+/b95ca3345b017460ce3cd54ed7bedf1f673f1e1e [formerly c3672cd3f7e2bd87d6de9dada499de82b62fae84]
https://android.googlesource.com/platform/frameworks/base/+/a82fabb86f72b8b41305e72b4f70e121549b4ecd [formerly 41725dedc33906aaafee36b2d6523596e2a8a52e]

Bug: 24967303
Bug: https://code.google.com/p/android/issues/detail?id=189710

Change-Id: I515e47bafcf39a2b1bdf92f11f623bef8fb6263c

Former-commit-id: 20ebb4c2d492b17fa8aec931e16305463de1f550