Keep freeform window surface translucent during resize.
Freeform window surfaces are translucent because most of the time they
have shadows. During a resize we stop displaying the shadow, which
might change the surface opacity from translucent to opaque. This change
only happens if the activity gets recreated during the resize, as this
triggers recalculation of the necessary opacity. If configuration change
happens the relayout will now contain new, opaque pixel format and cause
the recreation. The second blink will happen after we finish resizing,
as the surface needs to become translucent again.

Bug: 24668341
Change-Id: I450323276c49f176f0f6dfb3b21a5f6d742a8418

Former-commit-id: 2dfcf5ce2e96073d6cb875d048684e0a074ce2b6