Fix wrongly laid out navigation color view
The SystemUI visibility listener in DecorView
gets called between the measure and layout passes
and is therefore not allowed to change layout parameters.

This change makes sure that changes to the color view
layout parameters are applied eagerly when the insets
change instead of waiting for the views to become visible.

Bug: 24614374

Change-Id: If9df18f582163d0869c28a852c36697b1ce50621

Former-commit-id: fc7476da7a59f0fb7b07d6e188ae6d0b55f140b8