docs: Update code snippet so it doesn't depend on the sample app
The code snippet was copied straight from the MediaBrowserService
sample app. Since the code snippet consisted almost entirely of
calls to *other* methods in that sample app, the snippet wasn't
very helpful--and it wouldn't be helpful to add a lot of verbiage
explaining the sample app. So instead, I made the code more generic
(e.g. comments saying "You need to write code to check for XYZ")
followed by a link to the sample app.

Fixed other miscellaneous problems in the code snippets while we had
the file open.

See first comment for doc stage location.

bug: 20008288
Change-Id: Ic080ebffcc43e332ec3335b730c07ef5f8b64fa9

Former-commit-id: f01aca2caf43bedf35c7faa7401cbac45391cba1